Authentication guards

See merge request nox/b-saffer-dashboard!7

+++ apps/backend/src/controllers/project/project.controller.ts
@@ -8,18 +8,27 @@ import {
   Query,
   Delete,
   Req,
+  UnauthorizedException,
 } from '@nestjs/common';
 import { ProjectService } from 'src/services/project/project.service';
 import { CreateProjectDto } from '@bsaffer/api/project/dto/create-project.dto';
 import { UpdateProjectDto } from '@bsaffer/api/project/dto/update-project.dto';
-import { Public, Roles, KeycloakUser } from 'nest-keycloak-connect';
+import { SessionRequest } from 'src/auth/sessionData';
+import { isAdmin } from 'src/auth/methods/isAdmin';
 
 @Controller('project')
 export class ProjectController {
   constructor(private readonly projectService: ProjectService) {}
 
   @Post()
-  async create(@Body() createProjectDto: CreateProjectDto) {
+  async create(
+    @Body() createProjectDto: CreateProjectDto,
+    @Req() request: SessionRequest,
+  ) {
+    if (!isAdmin(request)) {
+      throw new UnauthorizedException();
+    }
+
     const project = await this.projectService
       .create(createProjectDto)
       .catch((error) => {
@@ -29,23 +38,19 @@ export class ProjectController {
     return project;
   }
 
-  //@Roles('Gebruiker')
   @Get()
-  async findAll(@Query('hidden') showHidden: string, @Req() req) {
-    console.log(req.session);
-
+  async findAll(
+    @Query('hidden') showHidden: string,
+    @Req() request: SessionRequest,
+  ) {
+    const admin = isAdmin(request);
     const projects = await this.projectService
-      .findAll(showHidden === 'true')
+      .findAll(admin ? showHidden === 'true' : false)
       .catch((error) => {
         console.error(error);
       });
 
     return projects;
-  }
-
-  @Get('user')
-  async Get() {
-    return KeycloakUser.name;
   }
 
   @Get(':id')
@@ -61,14 +66,23 @@ export class ProjectController {
   async update(
     @Param('id') id: string,
     @Body() updateProjectDto: UpdateProjectDto,
+    @Req() request: SessionRequest,
   ) {
+    if (!isAdmin(request)) {
+      throw new UnauthorizedException();
+    }
+
     return this.projectService.update(+id, updateProjectDto).catch((error) => {
       console.error(error);
     });
   }
 
   @Delete(':id')
-  async remove(@Param('id') id: string) {
+  async remove(@Param('id') id: string, @Req() request: SessionRequest) {
+    if (!isAdmin(request)) {
+      throw new UnauthorizedException();
+    }
+
     return this.projectService.remove(+id).catch((error) => {
       console.error(error);
     });
